% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brxsm_pla-data.R
\docType{data}
\name{brxsm_pla}
\alias{brxsm_pla}
\title{Small molecules from IVDr method}
\format{
A data frame with 41 plasmic metabolites concentration ranges
}
\usage{
data(brxsm_pla)
}
\description{
A dataset containing 41 metabolites from plasma
}
\keyword{datasets}
