% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNames.R
\name{cleanNames}
\alias{cleanNames}
\title{function to clean names for importation into databases}
\usage{
cleanNames(names)
}
\arguments{
\item{names}{\itemize{
\item a name or an array of
}}
}
\value{
clean name(s)
}
\description{
function to clean names for importation into databases
}
\examples{
originalID <- c("ddd.aaa", "ddd uuu", "ddd+aaa", "ddd*yyy", "ddd#dd", "ddd_fff")
originalID <-c(originalID, "ddd$ddd", "ddd@ddd", "dd_aa", "dad*")
cleanNames(originalID)
}
