% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-lipo-data.R
\docType{data}
\name{ext_lipo}
\alias{ext_lipo}
\title{Extended Lipoproteins from}
\format{
A data frame with 112 lipoproteins values
}
\usage{
data(lipo)
}
\description{
A dataset containing 112 Lipoproteins from IVDR method with additional extended parameters
\strong{calc} is from the summation of the raw values to find TL (total lipid) and
CH - FC to calculate CE.
\strong{pct} looks at subfractions considering the lipid composition, so what
portion does each of the raw V1TG, V1CH, V1PL contribute to the calculated V1TL.
\strong{frac} looks at subfractions considering the lipid distribution, so what
portion do each of V1TG, V2TG, V3TG, V4TG, V5TG contribute to VLTG
}
\details{
\itemize{
\item Fraction - lipoprotein fraction
\item calc - total lipids and Cholesterol esters
\item pct - sub-fraction lipid composition
\item frac - sub-fraction lipid distribution
}
}
\keyword{datasets}
