% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_lipo.R
\name{extend_lipo}
\alias{extend_lipo}
\title{extend_lipo
Calculates total lipids, fractions and percentages using data.frame "res"
from readLipo.r}
\usage{
extend_lipo(lipo)
}
\arguments{
\item{res}{Data frame with lipoprotein data for a single sample}
}
\value{
Data frame with new parameters.
\strong{calc} is from the summation of the raw values to find TL (total lipid) and
CH - FC to calculate CE.
\strong{pct} looks at subfractions considering the lipid composition, so what
portion does each of the raw V1TG, V1CH, V1PL contribute to the calculated V1TL.
\strong{frac} looks at subfractions considering the lipid distribution, so what
portion do each of V1TG, V2TG, V3TG, V4TG, V5TG contribute to VLTG
}
\description{
extend_lipo
Calculates total lipids, fractions and percentages using data.frame "res"
from readLipo.r
}
