% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_lipo.R
\name{extend_lipo_value}
\alias{extend_lipo_value}
\title{function to extend the lipo and return a vector of values}
\usage{
extend_lipo_value(lipo)
}
\arguments{
\item{lipo}{\itemize{
\item takes a lipo object
}}
}
\value{
a vector with values and no particular format
}
\description{
function to extend the lipo and return a vector of values
}
