% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLipoTable.R
\name{getLipoTable}
\alias{getLipoTable}
\title{get names and description of lipoproteins IVDr parameters}
\usage{
getLipoTable(extended = FALSE, withDensities = FALSE)
}
\arguments{
\item{extended}{if TRUE, the extended_lipo information is attached below the lipoTable}
}
\value{
a data.frame with information
}
\description{
get names and description of lipoproteins IVDr parameters
}
