% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPacsTable.R
\name{getPacsTable}
\alias{getPacsTable}
\title{This function is used to get the table of PACS data from the test data}
\usage{
getPacsTable()
}
\description{
This function is used to get the table of PACS data from the test data
}
