% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQcTable.R
\name{getQcTable}
\alias{getQcTable}
\title{Get a data frame with the QC table}
\usage{
getQcTable(matrixType = "SER", withValue = FALSE)
}
\arguments{
\item{matrixType}{A string with the matrix type. It can be "SER" or "URI"}

\item{withValue}{A boolean to include the value column}
}
\description{
Get a data frame with the QC table
}
