% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmTable.R
\name{getSmTable}
\alias{getSmTable}
\title{get names and description of brxsm plasma reports}
\usage{
getSmTable(matrixType = "SER")
}
\arguments{
\item{matrixType}{\itemize{
\item sample matrix type (URI, SER)
}}
}
\value{
a data.frame with information
}
\description{
get names and description of brxsm plasma reports
}
