% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipoWithDensities-data.R
\docType{data}
\name{lipoWithDensities}
\alias{lipoWithDensities}
\title{Lipoproteins with densities from IVDr method}
\format{
A data frame with 112 lipoproteins values
}
\usage{
data(lipoWithDensities)
}
\description{
A dataset containing 112 Lipoproteins
}
\details{
\itemize{
\item Fraction - lipoprotein fraction
}
}
\keyword{datasets}
