% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.parser-package.R
\docType{package}
\name{nmr.parser-package}
\alias{nmr.parser}
\alias{nmr.parser-package}
\title{nmr.parser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
