% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1r.R
\name{read1r}
\alias{read1r}
\title{read spectrum (processed)}
\usage{
read1r(file, numberOfPoints, nc = 0, endian = "little")
}
\arguments{
\item{numberOfPoints}{\itemize{
\item the number of points that should be read
}}

\item{nc}{\itemize{
\item power factor
}}

\item{endian}{\itemize{
\item endianness
}}

\item{filePath}{\itemize{
\item the path to the spectrum (processed)
}}
}
\value{
the spectrum
}
\description{
read spectrum (processed)
}
