% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEretic.r
\name{readEretic}
\alias{readEretic}
\title{extract xml quantification information from a bruker xml}
\usage{
readEretic(path)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}
}
\value{
the title
}
\description{
extract xml quantification information from a bruker xml
}
