% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLipo.r
\name{readLipo}
\alias{readLipo}
\title{extract lipoprotein  quantification information from a bruker xml}
\usage{
readLipo(path)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}
}
\value{
the values
}
\description{
extract lipoprotein  quantification information from a bruker xml
}
