% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPacs.R
\name{readPacs}
\alias{readPacs}
\title{extract pacs information from a bruker xml}
\usage{
readPacs(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder
}}
}
\value{
the values
}
\description{
extract pacs information from a bruker xml
}
