% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQc.R
\name{readQc}
\alias{readQc}
\title{read Qc info for importation into rolodex
extract qc information}
\usage{
readQc(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder (plamsa_qc_report.xml / urine_qc_report.xml)
}}
}
\value{
the values
}
\description{
read Qc info for importation into rolodex
extract qc information
}
