% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQuant.r
\name{readQuant}
\alias{readQuant}
\title{extract small molecules  quantification information from a bruker xml}
\usage{
readQuant(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder
}}
}
\value{
the values
}
\description{
extract small molecules  quantification information from a bruker xml
}
