% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLipo.r
\name{roldx_getLipo}
\alias{roldx_getLipo}
\title{get lipo for rolodex}
\usage{
roldx_getLipo(path)
}
\arguments{
\item{path}{\itemize{
\item path to the data
}}
}
\value{
a triad
}
\description{
get lipo for rolodex
}
