% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roldx_readLipo.R
\name{roldx_readLipo}
\alias{roldx_readLipo}
\title{get lipo for rolodex}
\usage{
roldx_readLipo(file)
}
\arguments{
\item{file}{\itemize{
\item path to the data
}}
}
\value{
a triad
}
\description{
get lipo for rolodex
}
