% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFolder.R
\name{scanFolder}
\alias{scanFolder}
\title{function to scan a folder and get a list of paths to parse}
\usage{
scanFolder(folder, options = list())
}
\arguments{
\item{folder}{\itemize{
\item the root folder to scan
}}

\item{options}{\itemize{
\item options to skip the prompt
}}
}
\value{
a list of paths
}
\description{
function to scan a folder and get a list of paths to parse
}
