% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brxsm_uri-data.R
\docType{data}
\name{brxsm_uri}
\alias{brxsm_uri}
\title{Small molecules from IVDr method}
\format{
A data frame with 26 urine metabolites concentration ranges
}
\usage{
data(brxsm_uri)
}
\description{
A dataset containing 26 metabolites from urine
}
\keyword{datasets}
