% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_lipo.R
\name{extend_lipo}
\alias{extend_lipo}
\title{extend_lipo
calculate the total lipids and its percentage using data.frame "res" from readLipo.r}
\usage{
extend_lipo(res)
}
\arguments{
\item{res}{\itemize{
\item data frame with lipoprotein data
}}
}
\value{
\itemize{
\item a data frame with new parameters
}
}
\description{
extend_lipo
calculate the total lipids and its percentage using data.frame "res" from readLipo.r
}
