% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.parser-package.R
\docType{package}
\name{nmr.parser-package}
\alias{nmr.parser}
\alias{nmr.parser-package}
\title{nmr.parser: reading what is required from NMR IVDr data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Julien Wist \email{julien.wist@murdoch.edu.au} (\href{https://orcid.org/0000-0002-3416-2572}{ORCID})

Authors:
\itemize{
  \item Reika Masuda \email{reika.masuda@murdoch.edu.au}
}

}
\keyword{internal}
