% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEreticF80.R
\name{readEreticF80}
\alias{readEreticF80}
\title{extract xml quantification information from a bruker xml}
\usage{
readEreticF80(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder
}}
}
\value{
the title
}
\description{
extract xml quantification information from a bruker xml
}
