% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExperiment.r
\name{readExperiment}
\alias{readExperiment}
\title{read experiment from a bruker folder (expno)}
\usage{
readExperiment(expname, opts = NULL)
}
\arguments{
\item{expname}{\itemize{
\item a path (or list of) to the expNo folder(s)
}}

\item{opts}{\itemize{
\item object with processing parameters, if TRUE reads from the file
what - choose what to read (acqus, procs, qc, title, eretic, spec, lipo, quant, pacs, all)
}}
}
\value{
a list with all read elements
}
\description{
read experiment from a bruker folder (expno)
}
