% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExperiment.r
\name{readExperiment}
\alias{readExperiment}
\title{read experiment from a bruker folder (expno)}
\usage{
readExperiment(
  file,
  what = c("acqus", "procs", "qc", "title", "eretic", "spec", "lipo", "quant", "pacs",
    "all", "specOnly"),
  options = list(specOpts = list(uncalibrate = FALSE, fromTo = c(-0.1, 10), length.out =
    44079))
)
}
\arguments{
\item{file}{\itemize{
\item the path to the expNo folder
}}

\item{what}{\itemize{
\item choose what to read (acqus, procs, qc, title, eretic, spec, lipo, quant, pacs, all)
}}

\item{options}{\itemize{
\item object with processing parameters, if TRUE reads from the file
}}
}
\value{
a list with all read elements
}
\description{
read experiment from a bruker folder (expno)
}
