% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParams.r
\name{readParams}
\alias{readParams}
\title{extract a parameter from a bruker file (procs or acqus)}
\usage{
readParams(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder
}}
}
\value{
the parameter
}
\description{
extract a parameter from a bruker file (procs or acqus)
}
