% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTitle.R
\name{readTitle}
\alias{readTitle}
\title{extract title from a bruker folder}
\usage{
readTitle(file)
}
\arguments{
\item{file}{\itemize{
\item the path to the expName folder
}}
}
\value{
the title
}
\description{
extract title from a bruker folder
}
